requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.27.1